requirements.txt
Django>=4.2
requests>=2.31.0
python-decouple>=3.8
python-dotenv>=1.0.0 